Complete refactoring of span drawers, heavily macro-ized to allow creation of translucent and additive span drawers. flatsize array altered to store enumeration value representing flat size.

+++ eternity/source/r_state.h
@@ -62,7 +62,7 @@ extern int *flattranslation;
 extern int *texturetranslation;
 
 // SoM: flat lump size
-extern int *flatsize;
+extern byte *flatsize;
 
 // Sprite....
 extern int firstspritelump;
